Removed / Backed Up Scripts
Removed scripts and placed a few in scripts/Backup
