+++ Assets/Scripts/Backup/NotesScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7656a617bf0b5147ace6151aa640770
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
