--- Assets/Scripts/CollectItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b7cc66ec33b29124abc9ab714e6a0d17
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
