--- Assets/Scripts/Dimensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fb169e00cfe30484fa8bb1da068d3bfd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
