--- Assets/Scripts/Elevator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 14381815e5367a943824a5e0c1b39b21
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
