--- Assets/Scripts/Finale.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d2b75189709b044f9a9e63957923802
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
