--- Assets/Scripts/Inventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e4d31fd765607f46af7c909a0240323
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
