--- Assets/Scripts/Minigame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c73936418258cb41a2e71b755d4598a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
