--- Assets/Scripts/Sounds.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 17349ba5f49314940998d49d9e9fc0a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
