--- Assets/Scripts/TeleportTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d3c4b0c4fb5606f49a046d32afdff28e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
