+++ Assets/StreamingAssets.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e6ce99eaa5168ce4d86c1e137cc60252
+guid: ae736e067c9278e469e2b50282de0bac
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
